set callbacks for led counters

+++ Core/Program/leds.c
@@ -43,7 +43,7 @@ typedef struct {
 } Led;
 
 typedef bool(*Leds_Predicate)(void);
-typedef uint32_t(*Leds_ValueGetter)(void);
+typedef uint8_t(*Leds_ValueGetter)(void);
 
 typedef struct {
 	Led led;
